added submit method with POST but client not working
